backend/src/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pdf2image
pymupdf
python-docx
pillow
spacy
transformers
torch
accelerate
sentencepiece
sacremoses
